Update splunk_hec.conf.tmpl
Added dir argument to disk buffer in splunk_hec_conf.tmpl

This change adds code to look for the SC4S_DEST_SPLUNK_HEC_DISKBUFF_DIR environmental variable and if present, it uses that directory for the directory of the disk buffer.

The dir() option allows you to change the directory used for disk buffers.  For 3.16 the feature is document at
https://www.syslog-ng.com/technical-documents/doc/syslog-ng-open-source-edition/3.16/administration-guide/disk-buffer

The caveat is if you have already encountered disk buffering, there will be a persist file in the previous location (default /var/lib/syslog-ng) and that persist file will prevent syslog-ng from sucessfully using the new location.