Fixes #273 make indexed fields optional
The fields sc4s_syslog_product and sc4s_syslog_format are not made optional in this change as the current use is intergral to the logic tree a future change could change the internal field name and conditional include the field in the json payload.

This change also adds new optional indexed fields for dest port when and container host.