feat(k8s): Provide Helm3 charts for sc4s (#1234)
This release replaces the original manifest-based deployment of sc4s with a new helm3 based deployment of sc4s with associated documentation. In addition, we are making one more change to the container registry configuration to clean up the package naming convention.

The helm chart features should be considered beta.