docs/sources/vendor/Checkpoint: Fix formatting
The formatting as previously given resulted in many of the template
tokens surrounded by brackets to be interpreted by the HTML convertor as
an HTML tag without a closing pair, leading the convertor dropping it.
Adding monospace modifiers fixes this issue.

Additionally, adjust the language for readability, and add some missing
links.