Update docker-systemd-general.md
For this statement:  "Make it a habit to check back here regularly to be sure any changes that may have been made to the template unit file below" is there a more concrete way to say this? I feel like the customer is going to be confused about how often they are expected to check back in. is it with every release?